requirements.txt
CherryPy==10.2.1
Flask==0.12.1
PasteScript==2.0.2
flask-sqlalchemy==2.3.2
